Make darker_theme_unittest.cc
